+++ app/webFrontend/src/app/unit/unit-form/free-text-unit-form/free-text-unit-editor/free-text-unit-editor.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit, ViewChild} from '@angular/core';
+import {AfterViewInit, Component, Input, ViewChild} from '@angular/core';
 import {FreeTextUnitCoreComponent} from '../../../free-text-unit/free-text-unit-core/free-text-unit-core.component';
 import {AceEditorComponent} from 'ng2-ace-editor';
 import 'brace';
@@ -7,13 +7,12 @@ import 'brace/theme/github';
 import {FormGroup} from '@angular/forms';
 import {UnitFormService} from '../../../../shared/services/unit-form.service';
 
-
 @Component({
   selector: 'app-free-text-unit-editor',
   templateUrl: './free-text-unit-editor.component.html',
   styleUrls: ['./free-text-unit-editor.component.scss']
 })
-export class FreeTextUnitEditorComponent implements OnInit {
+export class FreeTextUnitEditorComponent implements AfterViewInit {
 
   @Input() markdown: string;
 
@@ -25,10 +24,7 @@ export class FreeTextUnitEditorComponent implements OnInit {
   @ViewChild('editor')
   private editor: AceEditorComponent;
 
-  constructor(private unitFormService: UnitFormService) {
-  }
-
-  ngOnInit() {
+  constructor(public unitFormService: UnitFormService) {
   }
 
   ngAfterViewInit() {
@@ -36,7 +32,6 @@ export class FreeTextUnitEditorComponent implements OnInit {
       maxLines: 9999
     });
   }
-
 
   onTabChange($event: any) {
     if ($event.index === 1) {