+++ app/webFrontend/src/app/unit/unit.component.ts
@@ -3,6 +3,7 @@ import {IUnit} from '../../../../../shared/models/units/IUnit';
 import * as moment from 'moment';
 import {ActivatedRoute, Router} from '@angular/router';
 import {IUser} from '../../../../../shared/models/IUser';
+import {MessageService} from '../shared/services/message.service';
 import {ICourse} from '../../../../../shared/models/ICourse';
 
 @Component({
@@ -15,6 +16,7 @@ export class UnitComponent implements OnInit, AfterViewInit {
   @Input() units: IUnit[];
   unitId: string;
   users: IUser[] = [];
+  chatMessageCount: {[unitId: string]: number} = {};
 
   private getDeadlineDiffTime (deadline: string) {
     const momentDeadline = moment(deadline);
@@ -59,13 +61,18 @@ export class UnitComponent implements OnInit, AfterViewInit {
   }
 
   constructor(private route: ActivatedRoute,
-              private router: Router) {
+              private router: Router,
+              private msgService: MessageService) {
   }
 
   ngOnInit() {
     this.route.params.subscribe(params => {
       this.unitId = decodeURIComponent(params['unit']);
     });
+    Promise.all(this.units.map(async (unit) => {
+      const res = await this.msgService.getMessageCount({room: unit.chatRoom});
+      this.chatMessageCount[unit._id] = res.count;
+    }));
   }
 
   ngAfterViewInit(): void {