+++ CHANGELOG.md
@@ -15,12 +15,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Chat system access denial unit tests. [#989](https://github.com/geli-lms/geli/issues/989)
 - `TestHelper` class for shared API unit test functionality. [#989](https://github.com/geli-lms/geli/issues/989)
 - `extractSingleMongoId` variant of the `ExtractMongoId` utility function(s). [#989](https://github.com/geli-lms/geli/issues/989)
-- Styles for md-units [#867](https://github.com/geli-lms/geli/issues/867)
+- Show message count for unit chat. [#933](https://github.com/geli-lms/geli/issues/993)
+- Styles for free text units. [#867](https://github.com/geli-lms/geli/issues/867)
 - Export pdf with styled md units [#997](https://github.com/geli-lms/geli/issues/997)
+
 
 ### Changed
 - Update mongoose to 5.2.x. [#1004](https://github.com/geli-lms/geli/pull/1004)
 - Update contributors list. [#1007](https://github.com/geli-lms/geli/issues/1007)
+- Use `terser` instead of `uglify-js`. [#1018](https://github.com/geli-lms/geli/pull/1018) 
 
 ### Removed